#17 Classe para mapeamento dos registros criada
