+++ RASGraphApp/lib/nblibraries.properties
@@ -2,6 +2,16 @@ libs.CopyLibs.classpath=\
     ${base}/CopyLibs/org-netbeans-modules-java-j2seproject-copylibstask.jar
 libs.CopyLibs.displayName=Tarefa CopyLibs
 libs.CopyLibs.prop-version=2.0
+libs.Hibernate\ (JPA\ 2.0).classpath=\
+    ${base}/Hibernate (JPA 2.0)-3/antlr-2.7.7.jar;\
+    ${base}/Hibernate (JPA 2.0)-3/dom4j-1.6.1.jar;\
+    ${base}/Hibernate (JPA 2.0)-3/hibernate-commons-annotations-4.0.2.Final.jar;\
+    ${base}/Hibernate (JPA 2.0)-3/hibernate-core-4.2.15.Final.jar;\
+    ${base}/Hibernate (JPA 2.0)-3/hibernate-entitymanager-4.2.15.Final.jar;\
+    ${base}/Hibernate (JPA 2.0)-3/hibernate-jpa-2.0-api-1.0.1.Final.jar;\
+    ${base}/Hibernate (JPA 2.0)-3/javassist-3.18.1-GA.jar;\
+    ${base}/Hibernate (JPA 2.0)-3/jboss-logging-3.1.0.GA.jar;\
+    ${base}/Hibernate (JPA 2.0)-3/jboss-transaction-api_1.1_spec-1.0.1.Final.jar
 libs.junit.classpath=\
     ${base}/junit-2/junit-3.8.2.jar
 libs.junit.displayName=JUnit 3.8.2
@@ -16,3 +26,7 @@ libs.junit_4.javadoc=\
 libs.junit_4.prop-maven-dependencies=junit:junit:4.10:jar
 libs.junit_4.src=\
     ${base}/junit_4-2/junit-4.10-sources.jar
+libs.PostgreSQLDriver.classpath=\
+    ${base}/PostgreSQLDriver-3/postgresql-9.2-1002.jdbc4.jar
+libs.PostgreSQLDriver.displayName=Driver JDBC do PostgreSQL
+libs.PostgreSQLDriver.prop-maven-dependencies=postgresql:postgresql:9.2-1002.jdbc4:jar
