+++ RASGraphApp/nbproject/project.properties
@@ -1,6 +1,6 @@
 annotation.processing.enabled=true
 annotation.processing.enabled.in.editor=false
-annotation.processing.processor.options=
+annotation.processing.processor.options=-Aeclipselink.canonicalmodel.use_static_factory=false
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
@@ -29,7 +29,9 @@ dist.javadoc.dir=${dist.dir}/javadoc
 excludes=
 includes=**
 jar.compress=false
-javac.classpath=
+javac.classpath=\
+    ${libs.Hibernate (JPA 2.0).classpath}:\
+    ${libs.PostgreSQLDriver.classpath}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
